# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Clover Nuke Bot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  ok: boolean;
}

export interface Account {
  id: number;
  label: string;
  token: string;
  username?: string | null;
  discriminator?: string | null;
  avatarUrl?: string | null;
  balance: number;
  totalClaimed: number;
  totalTransferred: number;
  enabled: boolean;
}

export type AccountWithStatusConnectionStatus = typeof AccountWithStatusConnectionStatus[keyof typeof AccountWithStatusConnectionStatus];


export const AccountWithStatusConnectionStatus = {
  connected: 'connected',
  connecting: 'connecting',
  disconnected: 'disconnected',
  error: 'error',
} as const;

export type AccountWithStatus = Account & {
  connected: boolean;
  connectionStatus: AccountWithStatusConnectionStatus;
};

export interface AccountInput {
  label: string;
  token: string;
  enabled?: boolean;
}

export interface BotSettings {
  id: number;
  serverId: string;
  channelId: string;
  cloverId: string;
  cloverPrefix: string;
  nukeKeywords: string;
  giveCommand: string;
  claimDelayMin: number;
  claimDelayMax: number;
  enabled: boolean;
}

export interface BotSettingsInput {
  serverId: string;
  channelId: string;
  cloverId: string;
  cloverPrefix: string;
  nukeKeywords: string;
  giveCommand: string;
  claimDelayMin: number;
  claimDelayMax: number;
}

export type AccountStatusStatus = typeof AccountStatusStatus[keyof typeof AccountStatusStatus];


export const AccountStatusStatus = {
  connected: 'connected',
  connecting: 'connecting',
  disconnected: 'disconnected',
  error: 'error',
} as const;

export interface AccountStatus {
  accountId: number;
  label: string;
  username?: string | null;
  connected: boolean;
  status: AccountStatusStatus;
  balance: number;
  claimsThisSession: number;
  scrapThisSession: number;
}

export interface BotStatus {
  running: boolean;
  accounts: AccountStatus[];
  totalClaimsToday: number;
  totalScrapToday: number;
  uptime: number;
}

export interface TransferRequest {
  toUsername: string;
  amount?: number | null;
  accountIds?: number[] | null;
}

export type TransferResultResultsItem = {
  accountId: number;
  label: string;
  amount: number;
  success: boolean;
  error?: string | null;
};

export interface TransferResult {
  results: TransferResultResultsItem[];
  totalTransferred: number;
}

export interface TransferLog {
  id: number;
  fromAccountId: number;
  fromLabel?: string | null;
  toUsername: string;
  amount: number;
  success: boolean;
  error?: string | null;
  sentAt: string;
}

export type NukeEventWithClaimsClaimsItem = {
  accountId: number;
  label?: string | null;
  success: boolean;
  scrapGained: number;
  error?: string | null;
};

export interface NukeEventWithClaims {
  id: number;
  messageId: string;
  channelId: string;
  detectedAt: string;
  totalScrapClaimed: number;
  claimCount: number;
  claims: NukeEventWithClaimsClaimsItem[];
}

export type LogEntryLevel = typeof LogEntryLevel[keyof typeof LogEntryLevel];


export const LogEntryLevel = {
  info: 'info',
  warn: 'warn',
  error: 'error',
  debug: 'debug',
} as const;

export interface LogEntry {
  timestamp: string;
  level: LogEntryLevel;
  message: string;
  accountId?: number | null;
}

export type GetAccounts200 = {
  accounts: AccountWithStatus[];
};

export type GetTransfersParams = {
limit?: number;
};

export type GetTransfers200 = {
  transfers: TransferLog[];
};

export type GetEventsParams = {
limit?: number;
};

export type GetEvents200 = {
  events: NukeEventWithClaims[];
};

export type GetLogsParams = {
limit?: number;
};

export type GetLogs200 = {
  logs: LogEntry[];
};